backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
databases 
bcrypt 
python-jose[cryptography] 
python-dotenv 
mysql-connector-python
shapely
requests
matplotlib